SSCSSI-445: removed CITIZEN_LOGIN_SERVICE_V2
