+++ charts/sscs-tribunals-api/values.preview.template.yaml
@@ -37,7 +37,6 @@ java:
     ELASTIC_SEARCH_FEATURE: true
     ELINKS_V2_FEATURE_ENABLED: true
     DELETED_REDACTED_DOC_ENABLED: true
-    CITIZEN_LOGIN_SERVICE_V2: true
     ADD_LINK_TO_OTHER_ASSOCIATED_CASES_V2: true
     BYPASS_NOTIFICATIONS_SERVICE: true
     BYPASS_HEARING_API_SERVICE: true
