+++ src/main/resources/config/application.properties.old
@@ -141,7 +141,6 @@ feature.elasticsearch.enabled:  ${ELASTIC_SEARCH_FEATURE:false}
 feature.elinksV2.enabled: ${ELINKS_V2_FEATURE_ENABLED:false}
 feature.deleted-redacted-doc.enabled: ${DELETED_REDACTED_DOC_ENABLED:false}
 feature.hearing-recording-filter.enabled: ${UPLOAD_HEARING_FILTER_ENABLED:false}
-feature.citizen-login-service-v2.enabled: ${CITIZEN_LOGIN_SERVICE_V2:false}
 feature.add-link-to-other-associated-cases-v2.enabled: ${ADD_LINK_TO_OTHER_ASSOCIATED_CASES_V2:false}
 feature.log-selected-properties.enabled: ${LOG_SELECTED_ENVIRONMENT_VARIABLES:false}
 feature.hearings-recording-request.case-updateV2.enabled: ${HEARING_RECORDING_REQUEST_CASE_UPDATE_V2_ENABLED:false}
