+++ src/main/resources/config/application.yaml
@@ -243,8 +243,6 @@ feature:
         enabled: ${ADD_LINK_TO_OTHER_ASSOCIATED_CASES_V2:false}
     case-access-management:
         enabled: ${CASE_ACCESS_MANAGEMENT_FEATURE:false}
-    citizen-login-service-v2:
-        enabled: ${CITIZEN_LOGIN_SERVICE_V2:false}
     deleted-redacted-doc:
         enabled: ${DELETED_REDACTED_DOC_ENABLED:false}
     elasticsearch:
